requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Werkzeug
gunicorn
pytest  # <-- Adicione esta linha